🎯 Physical-First Architecture: Deterministic Centroids from Unicode Atoms
BREAKING CHANGE: Composition centroids now computed entirely in C++

This commit fundamentally shifts how token coordinates are determined:
Instead of projecting model embeddings to 4D coordinates, compositions
now derive their centroid as the AVERAGE of their constituent Unicode
atom coordinates. This makes coordinate assignment:

  ✓ Deterministic - same token always gets same coordinates
  ✓ Model-neutral - no dependency on embedding model
  ✓ Lossless - exact UTF-32 decomposition preserved
  ✓ Hierarchical - compositions are averages of atoms

Key Changes:
- projection_db.cpp: Complete rewrite for physical-first centroids
  • Parse token labels → UTF-32 codepoints
  • Compute atom coords via CoordinateMapper::map_codepoint()
  • Average to get composition centroid
  • Create composition_child entries linking to atoms
  • All computation in C++ - SQL is just the orchestrator

- Coordinate convention unified: CENTER at 2^31 (2147483648)
  • Atoms on 3-sphere surface (magnitude = 1)
  • Compositions inside sphere, scaling toward center with depth

- Hilbert indices cast to int64_t for PostgreSQL BIGINT compatibility

- Lanczos eigensolver improvements:
  • Deterministic starting vector (alternating pattern)
  • Skip null space eigenvector (λ₀ = 0)
  • Verbose CG solver diagnostics

- Added AUDIT_REPORT.md documenting current system state

Files: 15 modified, 2 added to archive

The token 'water' = average(w, a, t, e, r) atom coordinates 🌊

+++ cpp/src/seed_atoms_parallel.cpp
@@ -52,9 +52,10 @@ struct AtomRecord {
 // Generate atoms for a codepoint range
 void generate_range(uint32_t start, uint32_t end, std::vector<AtomRecord>& out) {
     out.reserve(end - start);
-    // NO normalization - store raw uint32 values as doubles
+    // COORDINATE CONVENTION: uint32 with CENTER at 2^31 = 2147483648
     // PostGIS GEOMETRY stores float64, which can exactly represent all uint32 values
     // (double has 53-bit mantissa, uint32 is 32-bit)
+    // Store uint32 coordinates DIRECTLY as double - NO reinterpretation as int32
     
     for (uint32_t cp = start; cp < end; ++cp) {
         if (cp >= constants::SURROGATE_START && cp <= constants::SURROGATE_END) {
@@ -69,20 +70,19 @@ void generate_range(uint32_t start, uint32_t end, std::vector<AtomRecord>& out) 
         rec.codepoint = static_cast<int32_t>(cp);
         rec.category = CoordinateMapper::categorize(cp);
         
-        // Store RAW 32-bit coordinates as signed int (LOSSLESS - same bit pattern as uint32)
-        // This is the authoritative coordinate storage
+        // Store uint32 coordinates as int32 (bit-preserving cast)
+        // This preserves the full 32-bit value - no information loss
         rec.coord_x = static_cast<int32_t>(coords.x);
         rec.coord_y = static_cast<int32_t>(coords.y);
         rec.coord_z = static_cast<int32_t>(coords.z);
         rec.coord_m = static_cast<int32_t>(coords.m);
         
-        // Store as CENTER-ORIGIN signed int32 (interpreted from uint32 bit pattern)
-        // This gives coordinates with 0 at center, ±2^31 at surface
-        // double can exactly represent all int32 values
-        rec.x = static_cast<double>(static_cast<int32_t>(coords.x));
-        rec.y = static_cast<double>(static_cast<int32_t>(coords.y));
-        rec.z = static_cast<double>(static_cast<int32_t>(coords.z));
-        rec.m = static_cast<double>(static_cast<int32_t>(coords.m));
+        // Store as double for PostGIS - DIRECT from uint32, no int32 reinterpretation
+        // This ensures CENTER (2^31) is stored as 2147483648.0, not as negative
+        rec.x = static_cast<double>(coords.x);
+        rec.y = static_cast<double>(coords.y);
+        rec.z = static_cast<double>(coords.z);
+        rec.m = static_cast<double>(coords.m);
         
         rec.hilbert_lo = static_cast<int64_t>(hilbert.lo);
         rec.hilbert_hi = static_cast<int64_t>(hilbert.hi);
